% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodes.R
\name{load_prodes}
\alias{load_prodes}
\title{PRODES - Deforestation Monitoring Project in the Legal Amazon by Satellite}
\usage{
load_prodes(
  dataset = "prodes",
  raw_data = FALSE,
  time_period,
  language = "eng"
)
}
\arguments{
\item{dataset}{A dataset name ("prodes").}
}
\value{
A \code{tibble} with the selected data.
}
\description{
Loads information on clearcut deforestation in the Legal Amazon and annual deforestation rates in the region.
}
\examples{
\dontrun{
# Download treated data (raw_data = FALSE) from 2010 (time_period = 2010)
# in portuguese (language = 'pt').
data <- load_prodes(
  raw_data = FALSE,
  time_period = 2010,
  language = "pt"
)
}

}
